requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4

# LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1
langchain-ollama>=0.1.0

# NLP & Embeddings
sentence-transformers>=2.2.0
sentencepiece>=0.1.99

# PDF & OCR
pytesseract>=0.3.10
opencv-python>=4.8.0
PyMuPDF>=1.23.0

# ML & Data Science
scikit-learn>=1.3.0
networkx>=3.0
matplotlib>=3.5.0

# Note: PyTorch with CUDA support should be installed separately:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
